**** 1 - new feature: The definitions of function members of templatized classes are now hooked up to their declarations.  If the function definition is seen before the instantiation request of the templatized class, then both are instantiated.  We do not yet implement the feature that the function definition can come after the instantiation request of the class template.  This is tested in Oink.
Added
  bool Env::currentScopeAboveTemplEncloses(Scope const *s)
which pops off the current scope if it is a template scope before
checking if currentScopeEncloses(); this is necessary to pretend that
the template scope doesn't exist when doing the typechecking for
function members of templatized classes in void Function::tcheck().

Also in void Function::tcheck() weakened the assertion
  xassert(s == qualifierScopes.top());
to be just that the two scopes' curCompounds match, with that of s
being at least as specific as the other; this is also necessary for
typechecking function members of templatized classes as the containing
scope's class is an instantiation of that of the qualifier scopes
top().

**** 2 - IMPORTANT new invariant:

We now maintain the invariant that the variable that is created by
typechecking the forward declaration of a function template or a
function member of a class template is re-used while typechecking the
definition of that function.  That is, when the first pass over the
Function object is made (with checkBody=false) we make sure that the
declaration's variable is re-used if this function was
forward-declared.  In this way we do not get two different declaration
variables made which a later analysis (cc_qual) would have to unify
(which is easy for function templates but seems hard for function
members of class templates.)  The tests for this feature are in Oink,
since a dataflow test is the easiest way to check that one variable
did not accidentally get split into two; therefore it is important to
run the Oink tests when checking in to Elsa to check that this
invariant continues to be maintained.

In the API to Env::instantiateTemplate() removed
  bool addTheInstV = true
and added
  Variable *funcFwdInstV=NULL
the purpose of this was that it still conveys the same boolean
information (with the current 'NULL' meaning the old 'true') but also
passes the variable which is the forward declaration of the function
template being instantiated.  This allows that forward variable to be
re-used when the first pass over the Function object is made (with
checkBody=false) so that we do not get two different declaration
variables made which a later analysis (cc_qual) would have to unify.

In D_name_tcheck() moved the converting of AST template arguments to
Semantic template arguments up to above the overload set resolution,
since the result of the former is used during the latter.  That is,
now after the overload set is found I look in it for the template
primary that fits the dt.type and then the instantiation of that
primary that best fits the semantics template arguments; This is then
set to the 'prior' which otherwise would have been NULL.  This allows
re-use of the declaration variable while typechecking the definition.

When we instantiate a class template we now instantiate all of its
function members as well that were defined separately.  Not sure if we
should do the superclass function members, but we don't for now.

**** 3 - desirable feature not implemented for now: The feature that
function templates and function members of templatized classes can use
different names for the template params of the declaration and
definition turned out to be hard and is currently not implemented.

Added test elsa/in/t0108b.cc which is t0108.cc but without using the
feature mentioned above.

**** 4 - other changes

void Env::provideDefForFuncTemplDecl() is now idempotent: it makes a
declaration look like a definition, but if it already has a funcDefn,
it leaves it (asserting its the same) rather than printing an error.

Made a new
  bool MatchTypes::match_Atomic(AtomicType *a, AtomicType *b, int matchDepth)
for use externally and which is also used to match the base types of
pointers to members.

In
  bool MatchTypes::match_TInfo(TemplateInfo *a, TemplateInfo *b, int matchDepth)
fixed a subtle bug where a specialization matched against its own
primary wouldn't match; now it does.

Fixed that t0108.cc did not compile since it had a bug: definition for
g() was named f() and members were not public.

Created
  void Env::insertBindings(Variable *baseV, SObjList<STemplateArgument> &sargs)
to unify into one place and delegate to the two different ways of
inserting bindings.  We were going to use it in two different places
since we were going to have to re-insert the bindings for function
template definitions where the names of the definition template params
differed from those of their declarations, but we backed off on
attempting to implement that feature.

Added
  void gdbScopeSeq(ScopeSeq &ss)
for printing scope sequences.

Added
  void Scope::gdb() const
that just delegates to Scope::desc()