Now, even for specializations (which didn't work in the last check-in), the definitions of function members of templatized classes are now hooked up to their declarations.  It still doesn't work if the definition comes after the instantiation.  This is tested in Oink.
This amounted to fixing the broken kludge in Env::lookupOneQualifier()
where Scott had special-cased specialization matching during lookup
that goes down through a PQ_qualifier by just looking through the
instantiations list of the relevant TemplateInfo for 1) type equality
of 2) the first argument.  I now do MM_ISO type matching on all the
arguments.  This code is factored out into the new
Env::getInstThatMatchesArgs().

The more-strict checking in this code found a bug: that Mutant
template instantiations were not being found by findMostSpecific()
when called in instantiateTemplate(), since they were being filtered
out of the search.  The invariant that I maintain now and that is
enforced by TemplateInfo::addInstantiation() and checked by
getInstThatMatchesArgs() is that no instantiation is repeated.  At the
call to addInstantiation() in instantiateTemplate() I now do a bit of
a kludge to avoid continuing on with typechecking the cloned AST: I
just throw it away, which jumps you over stage 5 of
instantiateTemplate().

The easy way to implement this was to just allow the duplicate mutant
to be created by instantiateTemplate and then just to filter it out.
This uses the dubious method of assigning to the reference argument,
but this is ameliorated by the fact that this assignment will only be
done if a boolean flag is set.

Also fixed a bug that in Declarator::mid_tcheck() the call to
env.getQualifierScopes(qualifierScopes, decl->getDeclaratorId()) was
coming before the decl->getDeclaratorId() had been typechecked.