Some ast visitors had a
        bool visitTemplateDeclaration(TemplateDeclaration *obj);

method which was careful to traverse the template instantiations; some
however did not.  There is now a class ASTTemplVisitor in the new file
cc_ast_aux.h inheriting from the generated ASTVisitor which contains
such a method.  All visitors now inherit from ASTTemplVisitor as for
each one it seemed to make sense to also traverse the template
instantiations.

The default is for only complete instantiations/specializations to be
traversed.  However one visitor, AmbiguityChecker, sets the flag in
ASTTemplVisitor which causes a traversal of all non-mutant members of
the instantiations list of the template primary, the partial
specializations, as well as the primary itself.  This seemed to make
sense given its semantics.