Started running oink on the 277K files in Hao's Red Hat 9 build interception which I'm calling ballA.  This is the first set of fixes. Some of these involve K and R mode, which is only available in oink, so all of the tests for these are in oink/test (listed below) and are driven by oink_test.mk.
        c_keyword_as_cpp_id.i
        double_goto.i
        gcc_Bool.i
        gcc_asm_as_func_qual.i
        gcc_assert_fail.i
        gcc_assert_fail2.i
        gcc_lack_of_asm_func_qual.i
        gcc_pretty.i
        kandr_call_undeclared_func.i
        kandr_func_def.i

---

Added that you can have an an assembler qualifier after the close paren
of a function declaration in gcc.
        extern void *tmpfile (void) __asm__ ("" "tmpfile64");

Added lexing "_Bool" the same as "bool".

Added declaring the builtin __assert_fail:
        void __assert_fail(char const *__assertion, char const *__file,
        unsigned int __line, char const *__function);

Added that in K and R C you can call a function that has not been
declared and it is assumed to be "int (...)".  For this, added a
LookupFlags:
        LF_IMPL_DECL_FUNC  = 0x100,   // K&R implicitly declare functions that are not found

Added that in gcc, __FUNCTION__ and __PRETTY_FUNCTION__ are synonymous
with __func__.  You had them parsing as strings.  Actually the are
variables just like __func__, so I treat them that way now.  Make a
test pass that didn't.

In
        void CFGEnv::addPendingGoto(StringRef name, S_goto *source)
I only add to the gotos map if the name is not already mapped.  This
is a hack until a real solution is found; there seems to be a
fundamental bug, which is that the goto graph is not invertible since
it is not one-to-one.