Two bugs revealed by "_Bool".
One was that the very first token was being lexed in C++ mode no
matter what.  Therefore the file "bool x;" was just fine in C mode,
since the very first token was lexed as C++; In contrast, "int y; bool
x;" would fail.  This is the kind of bad failure mode you get from a
non-orthogonality in the code: tokenFunc was being called in 2 places.

Another was that, whereas the semantics of "bool" are to be a bool in
C++ mode and an identifier in C mode, the semantics of "_Bool" are
exactly the reverse.  Therefore under no circumstance does it make
sense to simply lex them both the same way.  Now we simply introduce a
global typedef in the Env ctor when in C mode "typedef bool _Bool;".