Implemented the gcc feature that the 'then' part of a conditional expression may be omitted and if it is it means to evaluate to the value of the 'cond' part.
In oink this is tested to work for const evaluation and for $taint
analysis.

---

http://gcc.gnu.org/onlinedocs/gcc-3.4.1/gcc/Conditionals.html#Conditionals

Conditionals with Omitted Operands

The middle operand in a conditional expression may be omitted. Then if
the first operand is nonzero, its value is the value of the
conditional expression.

Therefore, the expression

     x ? : y

has the value of x if that is nonzero; otherwise, the value of y.

This example is perfectly equivalent to

     x ? x : y

In this simple case, the ability to omit the middle operand is not
especially useful. When it becomes useful is when the first operand
does, or may (if it is a macro argument), contain a side effect. Then
repeating the operand in the middle would perform the side effect
twice. Omitting the middle operand uses the value already computed
without the undesirable effects of recomputing it.