Neglected to check the two tests in with this; the implementation was checked in a minute ago.
Restrict is done more properly: CV_RESTRICT is now an official
CVFlags, inserted between volatile and owner.  Moved restrict away
from the odd place it was next to __attribute__ in gnu.gr and put it
in with "const" and "volatile" in the CVQualifier nonterminal; Did not
make it an UberModifier.  Added test in/gnu/d0076.cc for testing
parsing it: "restrict" is a C99-ism but right now we only lex
"__restrict__" as a gnu extension and I don't see a c99.gr; it should
be in in/c99.

Now parse qualifiers within the square brackets of an array as both a
DirectAbstractDeclarator and a DirectDeclarator.  Ben tells me that
this is a C99-ism and the qualifier is interpreted to go on the array
after it has degraded to a pointer.  Please see the test in
in/c99/d0077.cc.