Switched to using PtrMap most places.
NOTE: it is now really important that we map all the strings through StringTable before using them as a key to a map.

There is one test that seems to really heavily use the hashtable stuff: /home/ballA/postgresql-7.3.2-3/utf8_and_gb18030-mP23.i .  It improved a bit but much of its
time is still in the hashtable management code.