After doing binary operator overload resolution (unary is still not done), if resolution selects a built-in candidate, the E_binary node's type is computed according to the selected candidate.
This involves computing exactly which type the candidate returns,
which is nontrivial for polymorphic candidates.  For them, the
candidate return type is set to one of the new ST_PRET_* codes,
which denote an *algorithm* for computing the return type, given
the (concrete) parameter types.

I also puzzled through cppstd section 5 para 9, and also C99 section
6.3.1.8 para 1, to implement the "usual arithmetic conversions".