Subject: Re: oink tests find things elsa tests don't find      Date: Wed, 09 Mar 2005 17:48:49 -0800 (PST)      From: Scott McPeak <smcpeak@eecs.berkeley.edu>        To: "Daniel S. Wilkerson" <dsw@eecs.berkeley.edu>
> finding.  In particular, that in cc_scope.cc, STA_DEPEXPR was not handled
> in mangleSTemplateArgs().

+    case STemplateArgument::STA_DEPEXPR: // value-dependent expression
+      sb << "DEPEXPR-" << mangle(iter.data()->value.e->type);
+      break;

I think you want to print e, not just e's type.  A<N+3> and A<N+4> are
distinct types, but will both be mangled as A<DEPEXPR-int> the way the
code stands.

On the other hand, why would a non-concrete type need to be mangled?  Does
the linker imitator connect together (uninstantiated) template definitions
from multiple translation units?  (Maybe it should, maybe it shouldn't.)

        This is implemented.