Got rid of ASTTemplVisitor.  Recall that the idea was that ASTTemplVisitor inherited from ASTVisitor and helped client classes that inherited from it to traverse the lowered AST instead of the raw AST.  This was a brittle design and depended on the client classes knowing too much about what was going on and cooperating too much.
Now we use the new 'option dvisitor DelegatorASTVisitor' in the AST
language.  It generates an ASTVisitor that just delegates all the
visit and postvisit methods to a client class passed in to the dtor.
We then inherit from that to make LoweredASTVisitor which does all the
work of making sure that the client class is getting visit and
postvisit method calls for the lowered AST: 1) instantiated templates
but not template bodies, and 2) elaborated AST.

The client class should now use the following design pattern: 1)
privately inherit from ASTVisitor; 2) have a member LoweredASTVisitor
&loweredVisitor; 3) initialize its loweredVisitor(*this) at its ctor
call; 4) traversals should be launched as follows:
        Visitor vis;
        unit->traverse(vis.loweredVisitor);
The compiler will prevent you from saying unit->traverse(vis) since
the inheritance from ASTVisitor is private; 5) sub-traversals should be
launched as
        traverse(this->loweredVisitor)
unfortunately the compiler will not prevent you from launching them as
        traverse(*this) // WRONG!