In elsa created a "type printer for C" class TypePrinterC for real; the previous version was just an empty vessel.  The toCString() type printing functionality from the Type and AtomicType classes has been copied into TypePrinterC and augmented with functions where we roll our own virtual dispatch.
A global variable was created which the type printer sets to false
while it is running.
  bool global_mayUseTypeAndVarToCString = true;
The old toCString() methods now all have this line at the top.
  if (!global_mayUseTypeAndVarToCString) xfailure("suspended during TypePrinterC::print");
This caught a lot of bugs where the new printer was using some of the
old code.  I have left this mechanism in place for now to be removed
when it becomes inconvenient.

The idemtest script was augmented to run the output of pretty printing
through cpp.  This means that pretty printing idempotency now seems to
work on examples with templates in them, since the templates are
quoted with #if 0, #endif.  The idempotency test in/t0026.cc was
finally turned on again.