The print visitor now has #ifdef OINK conditionals where it needs in Oink to call the abstract value printer TypePrinterCO instead of the type printer TypePrinterC.  To make sure I didn't miss any spots I added a flag for turning off the type printer.  In Oink it is temporarily disabled in the TypePrinterCO ctor and enabled in the dtor.
  bool TypePrinterC::enabled = true;

There are a few places where the C type printer is needed even when
one is interested in printing values, such as printing out types in
template primaries which were never annotated with abstract values as
they aren't actually code.  The Restorer<bool> idiom is used to
temporarily enable the type printer in those cases.

  Restorer<bool> res0(TypePrinterC::enabled, true);