This check-in corresponds with the oink check-in marked:         STATIC SEPARATION OF TYPES AND VALUES
Changed the TypePrinter API to take a TypeLike, which is a typedef to
Type in Elsa and a superclass of Type and Value in Oink.  Added many
ifdefs based on the OINK preproc variable into cc_print.cc so that at
compile time the AST printer knows when to attempt to print an
abstract value annotating an AST node or Variable and when to just
print the type.

The identity of a type in Oink is a refinement of that in Elsa as we
have additional qualifiers.  This requires preventing two
optimizations in Elsa that rely on type identity not having a finer
granularity in an extension:

        1) Defeated the optimization in setCVQualifiers() where the
shallow clone is elided if the cv qualifiers do not change.  I
therefore factored the shallow Type clone out of setCVQualifiers()
into shallowCloneWithCV() so that in Oink when I override
setCVQualifiers() I can force the clone always and yet share the
shallow cloning code with Elsa.

        2) In BasicTypeFactory::makeCVAtomicType() prevent the re-use
of CVAtomicTypes wrapped around SimpleTypes that have CV_NONE
qualifiers.

Provided a means for Oink values and Elsa Types to be distinguished at
runtime from a void* pointer to one of them.  For this purpose added a
tag T_LAST_TYPE_TAG to the end of BaseType::Tag so that the Value tags
may be disjoin from the Type tags by starting the value tags at
T_LAST_TYPE_TAG+1.  This is really just documentation as I could have
used T_POINTERTOMEMBER as the last tag, but it would not have been
clear that I am doing that.

Did some renaming of identifiers to remove underscores.