Fixed enough bugs in the type serialization/de-serialization that I can turn on the full astxml_check3, both serializing AND now de-serializing Type annotations to the AST, and the first two tests now pass.  (Before I had the de-serializing part followed by the pretty print and diff turned off.)
Use the new 'xml' field modifier to the ast language; it means that
the field is to be serialized to/from xml; if a pointer just the
pointer is serialized and other code may handle serializing the object
pointed to, say if it is a Type annotation.

Now auto-generate three files that use the xml
serialization/de-serialization of the typesystem from a perl script
(that also contains the data): astxml_type_tok.pl; this makes it much
easier to modify the xml language!

Had two weaken the assertions in two places to allow for
de-serialization to pass 0 as the argument to ctors everywhere.  To
see these grep for XML_LOSS.

Added and then aborted the use of a <__Link from="pointer"
to="pointer"/> standalone tag to allow serialization of pointer-ness
independent of either of the objects that are pointing to one another.
This doesn't work but I left in the stand-alone tag mechanism as we
will probably want it later.  The code for processing the __Link tag
is commented out for now.

Turned off the mechanism for generating astxml_tokens.ids from
astxml_tokens.h as it doesn't seem to be being used.