Now first thirteen astxml_check3 (serialize and de-serialize types) work now.
Fixed bug that I was de-serializing SObjList-s and ObjList-s by making
the list and then inserting from the generic ASTList into it instead
of just using the list that was already embedded in the object (such
as the parameters list of a FunctionType).

Fixed bug that the code for hooking up a __Name tag and its containing
map was wrong.

Turned off assertion checking that you aren't putting a ReferenceType
into an array if the elt type of the ctor is NULL which it is during
de-serialization.

Fixed bug that we weren't creating EnumType::Value objects when we saw
the tag.

ArrayType's weren't rendering the newline after the '>' of the open
tag.