The class FullExpressionAnnot has always been a rather odd-man-out, so it is now officially an AST node.  This also greatly simplifies the serialization code.  During elaboration it annotates four AST nodes which now have it as a private member instead of a public one; it is hidden behind an accessor method which constructs it lazily.  Removed fullexp.h and its mention in the index.html.
I now use the new 'xml' field modifier that I just created in astgen.
So, now more fields are being serialized than before.

Added to and from xml methods for DeclaratorContext enums.

Had to make Function::cloneThunkSource point to a non-const Function
as otherwise the astgen code to serialize it would have to ask if it
is a pointer to const or not in order to be able to write out the
const cast at the traverse() call site; there is currently no nice way
to do that in astgen and it is a design decision to figure out how to
do that due to the use of strings to represent the user input.

Fixed the bug that during typechecking the D_array::size ast node was
being re-used for the size_E_new for an new expression which means the
AST was not a tree.  Now I clone it.

Fixed the bug that ints and longs were being serialized as one.  Fixed
the bug that signed and unsigned were being serialized as one.

Moved some of the generic xml helpers such as toXml_int() from
xml.cc/h to a new xmlhelp.c/h here in the ast repository.