Did what Scott asked here.
       PointerType::PointerType(CVFlags c, Type *a)
         : cv(c), atType(a)
       {
      -  // it makes no sense to stack reference operators underneath
      -  // other indirections (i.e. no ptr-to-ref, nor ref-to-ref)
      -  xassert(!a->isReference());
      +  // dsw: I had to put the 'if' here because the xml deserialization
      +  // code makes objects with NULL arguments and then fills them in
      +  // later
      +  if (a) {
      +    // it makes no sense to stack reference operators underneath
      +    // other indirections (i.e. no ptr-to-ref, nor ref-to-ref)
      +    xassert(!a->isReference());
      +  }
       }

    Rather than neutering the assertion for all uses of this constructor,
    make a new constructor.  Same goes for ReferenceType.  Don't change the
    semantics of the common case, make a new special case instead.

    What I recommend is to put a forward declaration in cc_type.h:

      class ReadXML;

    and then add constructor declarations like

      PointerType(ReadXML&);

    to cc_type.h.  Define them in cc_type_xml.cc.  The idea is that the
    class "ReadXML" serves to name this constructor and therefore document
    its role, even if that object is not actually used by the constructor.

I only needed to do it for Variable, Scope, and ArrayType.  I turned
the old assertions back on again.

I noticed that when I did it for ArrayType I could either make a new
makeArrayType() method on the type factory or I could special case it
and call the ctor directly and not go through the type factory.  I
have chosen to no longer go through the type factory for any of the
Type objects just for consistency.  I suppose this might be a problem
later.  Not sure what to do about it.