Previously we had added the 'item' field to the __Item tag but were not using it when parsing!  Added the 'item' field to the __Name tag as well and now for both tags are using it for parsing.  That is, when we see either tag, we look up the 'item' field in the hashtable and insert the element we find into the relevant list or map.
This immediate lookup is the first time that we are using the
hashtable during parsing instead of later during satisfyLinks().  Thus
we have rather a mixed strategy as the UnsatLink-s for normal object
fields are satisfied lazily rather than aggressively.

This design precludes the possibility of doing offline hashtable
construction: instead of maintaining an online hashtable, sorting the
UnsatLink objects after we are done generating them and then doing all
the lookups.  This design also means that there is now an order
dependency: you had better make an object with an .id field of 0xabc
before you reference it in a list or map.

However it also prevents us from making lots of UnsatLink objects and
inserting them into the List and then looking them up later.

This design also prevents us from checking that the correct kind of
thing is being inserted onto a list.