In order to implement xml serialization of classes with multiple inheritance (in Elsa, CompoundType) some global changes had to be made to the XmlReaderManager.
Whenever we generate an UnsatLink, we have to record the type of the
link to be satisfied.  That is, if we de-serialize an object that has
a field
        Foo *foo1;
Then when record the unsatisfied link to foo1, we need to record also
that it is a link to a Foo*.  This is necessary because when we
satisfy the link, it may be with a subclass of the desired class.
With only single inheritance, a reinterpret cast of the pointer will
work, but with multiple inheritance, the cast has to be done correctly
within the type system as the value of the pointer changes.

Note that we were generating code that recorded the kind before for
List classes for another purpose.  Now we record the kind for all
classes.

This process necessitates the addition of two more
"disjunctive-virtual-dispatch" functions: such functions are a design
pattern I use in the de-serializer where each XmlReaderManager
subclass registered with the XmlReaderManager is tried in turn to see
which one can handle a given tag for a given function.  The function
either handles it and signals that it did, or doesn't and signals that
it didn't.  If all fail, then it is an error.

In this case the new functions are these two:

  virtual bool upcastToWantedType(void *obj, int kind, void **target, int targetKind) = 0;

This function does the up-cast to deal with the mismatch between the
type of the de-serialized object that is going to satisfy a link and
the type of the link.

  virtual bool recordKind(int kind, bool& answer) = 0;

This function answers the question: for this kind (of tag), do we need
to record the map from its id to its kind in the new id2kind at the
time when we record the map from its id to is instantiation in id2obj?

This is purely an optimization: we could record them all.  However
they are only needed for those which are going to be involved with
multiple inheritance so that we know which subclass we are casting out
of when we satisfy the link, as mentioned above.

In the case of the AST, there is no multiple inheritance, so both of
these functions must exist and simply say 1) that there is nothing
interesting to record and 2) if you try to do a non-trivial up-cast it
is an error.  The implementation of the generated functions is
non-trivial however, as due to the nature of the
disjunctive-virtual-dispatch they must have this behavior only on
recognized AST classes; on unrecognized classes they should simply
pass on doing anything.

In the case of the typesystem where CompoundType multiply inherits,
implementations of recordKind() and upcastToWantedType() had to be
provided for TypeXmlReader.

Additionally, agreeable to the system mentioned above, the identity of
CompoundType is managed very carefully in the identity functions
idPrefix(), addr(), and printed(), which all work together.  Note that
the generic function for returning the address to be used as an id,
addr(), had to changed to work with the other two functions mentioned
as it is another form of identity.  This was a bug in the case of the
superclasses of CompoundType.

The ul() macro that created UnsatLink-s for nodes (not ints, not
lists) had to be changed to also record the kind of the node; this
required adding more actual information at the call site, which was
done by returning to the class definitions being de-serialized an
examining the field types of each field in question.  It is no longer
legal to make an UnsatLink without a kind.

In order to have a kind for some of the fields, new tokens had to be
introduced: XTOK_Type, XTOK_AtomicType and XTOK_NamedAtomicType.
These are never explicitly serialized, but are used for this
kind-recording purpose.  That is, the notion of a kind has been
abstracted to the point of separation from the literal token set.
Such is the nature of engineering.

Almost all of the type-serialization-deserialization tests in
checkxml.mk have been turned on except one due to some bug with
serializing or typechecking an E_charLit.

In addition, fixed a bug that we needed a special-purpose ctor
PointerToMemberType(ReadXML&) as the normal one is making assertions
that its arguments are not null which fails during deserialization.