In AST: Modified the ast language as follows:
If you want something serialized as xml, you have to indicate its "identity class"

    1. "xml" for scalars

    2. "xml_CLASS" for annotations that are pointers and should use
    CLASS followed by heap address as their identity string;

    We assume pointer annotations in ctor arguments must be serialized;
    there is no place to put a modifier.  In this cases we default
    CLASS to "TY".  Hopefully this will be fixed in the future.

If you serialize something that is embedded and not a pointer, you
have to indicate that by using the modifier "xmlEmbed_KIND" where KIND
is "List" or "NameMap".

In Elsa: Turned on serialization of these two fields.  This gets all
of the 714 xml type serialization tests to pass.

  -> PQ_qualifier {
       public ObjList<STemplateArgument> sargs;
     }

  -> PQ_template {
       public ObjList<STemplateArgument> sargs;
     }