Incorporated Oink's mangledName() and fullyQualifiedMangledName() as new methods on Variable, removing them from Oink.  They have had '0' appended to their names to make sure we don't collide with any existing functions such as the one in Scope.  When you change the semantics, change the name.
We assert, per Scott's suggestion, that fullyQualifiedMangledName0()
and mangledName0() are only called on linker-visible Variables.

Added an xml field to Variable: fullyQualifiedMangledName.  This is
essentially the name the linker will use to identify variables across
translation units.  It is intended to resemble somewhat the output of
nm -C.  Upon de-serialization it is discarded as it is redundant once
the Variable and its context have been fully de-serialized.

I inlined fqn_STemplateArgs() and Scope::fullyQualifiedName() with
mangle/doMangle=false to simplify things and because I didn't need the
other behavior.  mangleSTemplateArgs() is therefore no longer used so
I removed it.

In a C program each TranslationUnit is treated as if the entire thing
were wrapped in an 'extern "C" {}' block.  This means that the globals
in C programs also have the flag DF_EXTERN_C.