Added most of the mechanism for de-serializing source locations. Source locations are maintained in a highly optimized form by a SourceLocManager.  Therefore de-serializing them involves having a SourceLocManager in an initialized state.  This requires serializing and de-serializing the SourceLocManager.  This amounts to writing out the line lengths of each file.
However that requires an extension to the xml format so that we can
write and read blocks of numbers as a list of data (as stand-alone
tags would be just too verbose even for xml).  This was done in a way
conforming to the way you would do it so that the xml schema language
could say "this is a list of numbers"; namely, the list is simply
whitespace separated.

We serialize and de-serialize the data, but we do not yet use it to
initialize the state of the SourceLocManager.  Specifically we do the
following.

In smbase, added methods for the serialization code to get at the
internals of the SourceLocManager and SourceLocManager::File classes.
I did not want 'friend' references going across repositories, so I
couldn't think of any other way.

We generalized the xml lexer so that any alphanum token is a word, not
just those starting with an alpha or dot.  This is so we can parse
numbers as word tokens.

Added the <File> and <LineLengths> tags to hold metadata on files and
the list of their lengths respectively.  Note that files and
translation units are in a many-many relationship to one another.

The File tags all come before the TranslationUnit tag so that the
SourceLocManager can be initialized before the incoming SourceLoc-s.

The function parseOneTag() was generalized to parseOneTagOrDatum() as
it now sometimes encounters data.  When it does it calls the new
XmlReaderManager::registerStringToken() to register it.  Only the new
FileXmlReader has a non-trivial implementation of this function and
only for the LineLengths tag/object.