Added the new xmlAttrQuote()/xmlAttrDeQuote() to xmlhelp.cc. xmlPrintPointer() originally added by Scott is now moved from asthelp.h to xmlhelp.cc.
In the generated code we no longer generate in the xml visitor code an
inclusion of strutil.h as we no longer use quoted() and
parseQuotedString() to quote/unquote XML attributes, since, as Scott
points out, that's not what the spec says to do.  Instead we now use
the new xmlAttrQuote() and xmlAttrDeQuote().

A few years ago there was a first attempt at XML serialization
imitating the design of debug print.  This was abandoned but never
removed from the code.  I have now removed it.  example.ast no longer
requests the old xml system but the new one.