(As in the code generated by astgen) We no longer use quoted() and parseQuotedString() to quote/unquote XML attributes, since, as Scott points out, that's not what the spec says to do.  Instead we now use the new xmlAttrQuote() and xmlAttrDeQuote().
We no longer need a special quoted_printXml() for printing DeclFlags
as all the attribute values are now quoted, including pointers.  There
are two kinds of attribute values, 1) those that come from user data
and could contain any character, and 2) those that are generated
internally as a representation of an internal data-structure, such as
a stringification of an enum value.  Those of type 1) had always been
quoted.  I had managed to avoid quoting all those of type 2) as it is
rather straightforward to simply avoid putting strange characters like
double-quotes into the external representation of an internal
structure.  However, Scott decided that it would be cool to change
that when the DF_EXTERN_C flag was changed to be represented by
'(extern "C")'.  Scott had done this with a one-off change, but this
is dangerous and ad-hoc so for uniformity I changed *all* attributes
to be quoted.  This is probably way more inefficient, but it works and
prevents future bugs.