Added an "xml_verbatim" section that is code inserted into the generated xml serialization class.
Added the shouldSerialize() feature: if a member of a .ast file is
marked xmlShouldSerialize then the shouldSerialize() method in the
serialization class is consulted before a pointer to this member or
the member itself are serialized.  This allows pruning of the
serialization while at the same time maintaining a consistent
serialization where we only have serialized pointers to serialized
extant things (no dangling pointers).

Fixed a bug where pointers to empty AST lists were being serialized
even if the list was empty and was therefore not being serialized.