In ast: Added an "xml_verbatim" section that is code inserted into the generated xml serialization class.
Added the shouldSerialize() feature: if a member of a .ast file is
marked xmlShouldSerialize then the shouldSerialize() method in the
serialization class is consulted before a pointer to this member or
the member itself are serialized.  This allows pruning of the
serialization while at the same time maintaining a consistent
serialization where we only have serialized pointers to serialized
extant things (no dangling pointers).

Fixed a bug where pointers to empty AST lists were being serialized
even if the list was empty and was therefore not being serialized.

In elsa: Use the xml_verbatim and shouldSerialize xml annotations in
the .ast files to filter the serialization of Variables.

If the tracing flag no-danglingPointers is passed, then during
de-serialization if there is a pointer to an object that was not
serialized (a dangling pointer) then there is an assertion failure.

Turned on this -tr no-danglingPointers check for the xml serialization
tests when both the AST and Types are serialized (not on for just the
AST serialization as I didn't want to add mechanism to filter out all
the types; the AST tests are just there to catch any AST serialization
failures before running what I consider to be the real tests, the
AST+Types serialization).

Removed the special-case treatment of Variable members; the generic
shouldSerialize() mechanism now handles that.  Duh: you get
polymorphism over containers of different kinds of things by using
macros if the macros all call a function with the same name on the
contained elements as long as that function is overloaded for every
single one of them; took a while to think of that design pattern.

Now: 1) all Variable fields added to the ast have to have the
xmlShouldSerialize attribute, and 2) we have to manually call
shouldSerialize (by using the PRINT_ANNOT_MAYBE() macro) in
xml_type_writer.cc.  If either of these are omitted it is an error.
This will be detected more thoroughly when xml idempotency tests are
added.

Xml serialization should now be idempotent so the xml serialization
tests should now also have a version where the diff the xml rather
than the output of the debug printer.  Have not written these tests
yet.

There are some known bugs: BaseClassSubobj::parents list members are
not being serialized a second time (after de-serialization and
re-serialization); this does not affect C++ so # I leave it for now; I
looked at this for a while and could not find a good way to debug it.
It is very hard to debug the de-serialization code as it has many
passes and breaks the type system everywhere.